Adding and setup localization, fixing component and style bugs
