+++ package.json
@@ -18,7 +18,7 @@
     "react-router-dom": "6.2.2",
     "react-scripts": "5.0.0",
     "redux": "4.1.2",
-    "redux-thunk": "^2.4.1",
+    "redux-thunk": "2.4.1",
     "typescript": "4.6.2",
     "web-vitals": "2.1.4"
   },
