+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { Suspense } from 'react'
 
-import Layout from './components/navigation/Layout'
+import Layout from './components/layout/Layout'
 
 function App(): JSX.Element {
   return (
