+++ src/components/button/Button.module.scss
@@ -1,7 +1,7 @@
 @import '../../styles/colors.scss';
 
 .button {
-  font-size: 20px;
+  font-size: 25px;
   width: 200px;
   padding: 15px 0;
   text-align: center;
@@ -13,6 +13,9 @@
   cursor: pointer;
   position: relative;
   overflow: hidden;
+  p {
+    position: relative;
+  }
   .span {
     background: $blue;
     height: 100%;
@@ -21,7 +24,6 @@
     position: absolute;
     left: 0;
     bottom: 0;
-    z-index: -1;
     transition: 0.5s;
   }
   &:hover
