+++ src/components/layout/Layout.module.scss
@@ -1,0 +1,19 @@
+@import '../../styles/fonts.scss';
+@import '../../styles/colors.scss';
+@import '../../styles/images.scss';
+
+
+* {
+  margin: 0;
+  padding: 0;
+  font-family: $Creepster, $Neucha;
+}
+.banner {
+  background-image: $dark-gradient, $background;
+  background-size: cover;
+  background-position: center; 
+  background-attachment: fixed;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+}
