--- src/components/navigation/Layout.module.scss
@@ -1,14 +1,0 @@
-@import '../../styles/fonts.scss';
-
-* {
-  margin: 0;
-  padding: 0;
-  font-family: $rnm-font, cursive;
-}
-.banner {
-  background-image:linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url(../../img/background.jpg);
-  background-size: cover;
-  background-position: center; 
-  background-attachment: fixed;
-  height: 100vh;
-}
