+++ src/config/i18n.ts
@@ -1,0 +1,28 @@
+import i18n from 'i18next'
+import { initReactI18next } from 'react-i18next'
+
+import translationEN from './locales/en/translation.json'
+import translationRU from './locales/ru/translation.json'
+
+const resources = {
+  en: {
+    translation: translationEN,
+  },
+  ru: {
+    translation: translationRU,
+  },
+}
+
+i18n.use(initReactI18next).init({
+  resources,
+  fallbackLng: 'en',
+  debug: false,
+
+  interpolation: {
+    escapeValue: false,
+  },
+})
+
+i18n.t('')
+
+export default i18n
