--- src/pages/character/Character.module.scss
@@ -1,13 +1,0 @@
-@import '../../styles/colors.scss';
-
-.content {
-  width: 100%;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-80%);
-  text-align: center;
-  color: $blue;
-  h1 {
-    font-size: 90px;
-  }
-}
