--- src/pages/character/Character.tsx
@@ -1,26 +1,0 @@
-import React from 'react'
-import { useNavigate } from 'react-router-dom'
-import { useTranslation } from 'react-i18next'
-import '../../utils/i18n'
-
-import Button from '../../components/button/Button'
-
-import style from './Character.module.scss'
-
-function CharacterPage(): JSX.Element {
-  const { t } = useTranslation('translation')
-  const navigate = useNavigate()
-  return (
-    <div className={style.content}>
-      <h1>
-        {t('character.title')} <p>{t('character.page')}</p>
-      </h1>
-      <Button
-        label={t('character.home')}
-        onClick={(): void => navigate('/home')}
-      />
-    </div>
-  )
-}
-
-export default CharacterPage
