+++ src/pages/main/Main.tsx
@@ -1,23 +1,34 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
+import { useTranslation } from 'react-i18next'
+import '../../config/i18n'
 
 import Button from '../../components/button/Button'
 import style from './Main.module.scss'
 
 function MainPage(): JSX.Element {
+  const { t } = useTranslation()
   const navigate = useNavigate()
 
   return (
     <div className={style.content}>
       <h1>
-        Rick and Morty <p>Explorer</p>
+        {t('title')} <p>{t('main.page')}</p>
       </h1>
-      <Button
-        label="characters"
-        onClick={(): void => navigate('/characters')}
-      />
-      <Button label="locations" onClick={(): void => navigate('/locations')} />
-      <Button label="episodes" onClick={(): void => navigate('/episodes')} />
+      <div className={style.buttons}>
+        <Button
+          label={t('main.characterButton')}
+          onClick={(): void => navigate('/characters')}
+        />
+        <Button
+          label={t('main.locationButton')}
+          onClick={(): void => navigate('/locations')}
+        />
+        <Button
+          label={t('main.episodeButton')}
+          onClick={(): void => navigate('/episodes')}
+        />
+      </div>
     </div>
   )
 }
