+++ src/styles/colors.scss
@@ -1,2 +1,3 @@
 $blue: #4AADC9;
 $white: #fff;
+$dark-gradient: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5));
