+++ src/styles/fonts.scss
@@ -1,3 +1,5 @@
-@import url('https://fonts.googleapis.com/css2?family=Creepster&display=swap');
+// @import url('https://fonts.googleapis.com/css2?family=Creepster&display=swap');
+@import './fonts/stylesheet.scss';
 
-$rnm-font: 'Creepster';
+$Creepster: 'Creepster';
+$Neucha: 'Neucha';
