+++ src/styles/images.scss
@@ -1,0 +1,1 @@
+$background: url(../../img/background.jpg);
