--- src/utils/i18n.ts
@@ -1,15 +1,0 @@
-import i18n from 'i18next'
-import { initReactI18next } from 'react-i18next'
-
-i18n.use(initReactI18next).init({
-  fallbackLng: 'en',
-  debug: false,
-
-  interpolation: {
-    escapeValue: false,
-  },
-})
-
-i18n.t('character')
-
-export default i18n
